Read in ECG and EGM data from LabSystem Pro

This function allows for reading in LS Pro data based on their text export of signals. Signals can be exported directly from the LS Pro system.

The LabSystem Pro was acquired by Boston Scientific from the original company Bard. They are a common electrophysiology signal processing device for visualization and measurement of intracardiac signals.

Usage

read_lspro(file, n = Inf)

read_lspro_header(file)

read_lspro_signal(file, n = Inf)

Arguments

file

The path to the file where the data is located. It must be a *.txt file. See details below about its format.

n

Number of signal values to return (this will be the same for each channel of data). Defaults to all values.

Value

An egm class object that is a list of eps signals the format of a data.table, with an attached header attribute that contains additional recording data.

Data Export

The steps to data export are as follows.

Start LabSystem PRO

Open a patient record

Display a waveform recording in a Review Window

Scroll to a point of interest in a waveform recording

Right click on the review window to the left of the region of interest

Select an Export option, either a default time range or the entire visible page (which depends on the sweep speed).

Data Format

[Header] Recording info – contains (example):

[Header]<CR><LF>
     File Type: 1<CR><LF>
     Version: 1<CR><LF>
     Channels exported: 22<CR><LF>
     Samples per channel: 5000<CR><LF>
     Start time:  6:55:24<CR><LF>
     End time:  6:55:29<CR><LF>
     Ch. Info. Pointer: 320<CR><LF>
     Stamp Data: T<CR><LF>
     Mux format: 0<CR><LF>
     Mux Block Size: <CR><LF>
     Data Format 1<CR><LF>
     Sample Rate: 1000Hz<CR><LF>

[Header] Channel info (per channel example):

  Channel #: 1<CR><LF>
  Label: III<CR><LF>
     Range: 5mv <CR><LF>
     Low: 1Hz<CR><LF>
     High: 100Hz<CR><LF>
     Sample rate: 1000Hz<CR><LF>
     Color: 0000FF<CR><LF>
     Scale: -7<CR><LF>

[Data] As described below:

    -256,-1056,576,-256,320,-736,144,576,-592,176,608,240,176,-560,496,-
    144,0,0,-32,-48,-32,-80<CR><LF>

Channel Data is interleaved in the example above (sample indexed at 1):

1	2	3	...	22

Ch1:1	Ch2:1	Ch3:1	...	Ch22:1

Ch1:2	Ch2:2	Ch3:2	...	Ch22:2

Ch1:3	Ch2:3	Ch3:3	...	Ch22:3

...	...	...	...	...

Ch1:5000	Ch2:5000	Ch3:5000	...	Ch22:5000