Identify the color for a channel based on palettes

This primarily restricts the colors to color-space safe options. It is intended to be used with add_colors() to provide a color scheme for the ggm object. It has been exposed to users for custom or advanced theming options.

Usage

color_channels(x, palette, mode = "dark")

Arguments

x

Vector of character names of requested ECG or EGM leads

palette

A character choice from the below options that describe the color choices to be used for plotting. If set to the default, which is NULL, no changes to the colors for individual channels will be performed. If a positive choice is made, then the background mode argument will be set to dark as the default, unless otherwise specified. WARNING: This is an experimental argument, and may be moved in future version.

NULL: no changes to the colors will be made. DEFAULT.

material: a colorscheme based off of the Material Design color scheme

mode

A character string from c("dark", "light") to describe the base/background color settings to be used. If there are preset channel colors that were exported in the egm object, these colors will be used for the individual channels. If palette is specified, then the dark option will be set automatically (a palette choice cannot be made without understanding the background to plate it across). WARNING: This is an experimental argument, and may be moved in future version.

The dark theme mimics the "white on black" scheme seen in LabSystem Pro format (and most other high-contrast visualizations), for minimizing eye strain. This calls the theme_egm_dark() function. DEFAULT.

The light theme mimics the "black on white" colors seen in the Prucka system.

NULL removes any theme, and uses the default ggplot2::ggplot() settings

Value

Vector of hex code colors as character based on the selected palette and light/dark mode